feat(atendimento): ajuste - update-delete

+++ LABSchool-Manager/src/app/Atendimentos/components/edit/edit.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AtendimentoService } from 'src/app/services/atendimento.service';
-import { Atendimento } from '../../model/atendimento.moel';
+import { Atendimento } from '../../model/atendimento.moel'; // Notei que havia um erro de digitação no nome do arquivo ".moel"
+import { forkJoin, throwError } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 
 @Component({
   selector: 'app-edit',
@@ -10,7 +12,14 @@ import { Atendimento } from '../../model/atendimento.moel';
 })
 export class EditComponent implements OnInit {
 
-  atendimento: any = [];
+  atendimento: Atendimento = {
+    id: undefined,
+    dataHora: '',
+    descricao: '',
+    statusAtivo: false,
+    alunoId: 0,
+    pedagogoId: 0
+  };
   alunos: any[] = [];
   pedagogos: any[] = [];
   formInvalid: boolean = false;
@@ -23,28 +32,42 @@ export class EditComponent implements OnInit {
 
   ngOnInit() {
     const atendimentoId = +this.route.snapshot.params['id'];
-    this.getAtendimento(atendimentoId);
-    this.getAlunos();
-    this.getPedagogos();
+    forkJoin([
+        this.atendimentoService.getAlunos(),
+        this.atendimentoService.getPedagogos()
+    ]).subscribe(() => {
+        this.getAtendimento(atendimentoId);
+    });
   }
 
   getAtendimento(id: number): void {
-    this.atendimentoService.getAtendimento(id).subscribe(
-      (atendimento: Atendimento) => {
-        if (atendimento) {
-          this.atendimento = atendimento;
-        } else {
-          console.error('Atendimento não encontrado.');
+    this.atendimentoService.getAtendimento(id).pipe(catchError(this.handleError)).subscribe(
+        (response: any) => {
+            if (response && response.atendimento) {
+                const atendimento = response.atendimento;
+                atendimento.dataHora = this.convertToDateInputFormat(atendimento.dataHora);
+                this.atendimento = atendimento;
+            } else {
+                console.error('Atendimento não encontrado.');
+            }
+        },
+        (error: any) => {
+            console.error('Erro ao obter o atendimento', error);
         }
-      },
-      (error: any) => {
-        console.error('Erro ao obter o atendimento', error);
-      }
     );
   }
 
+  convertToDateInputFormat(dateStr: string): string {
+    if (dateStr.includes('-')) { // Se a data já estiver no formato YYYY-MM-DD
+        return dateStr;
+    }
+    const parts = dateStr.split('/');
+    return `${parts[2]}-${parts[1].padStart(2, '0')}-${parts[0].padStart(2, '0')}`;
+}
+
+
   getAlunos(): void {
-    this.atendimentoService.getAlunos().subscribe(
+    this.atendimentoService.getAlunos().pipe(catchError(this.handleError)).subscribe(
       (alunos: any[]) => {
         this.alunos = alunos;
       },
@@ -55,7 +78,7 @@ export class EditComponent implements OnInit {
   }
 
   getPedagogos(): void {
-    this.atendimentoService.getPedagogos().subscribe(
+    this.atendimentoService.getPedagogos().pipe(catchError(this.handleError)).subscribe(
       (pedagogos: any[]) => {
         this.pedagogos = pedagogos;
       },
@@ -64,13 +87,12 @@ export class EditComponent implements OnInit {
       }
     );
   }
-
   salvarAtendimento(): void {
-    if (!this.atendimento.descricao || !this.atendimento.alunoId || !this.atendimento.pedagogoId) {
+    if (!this.atendimento.descricao || !this.atendimento.alunoId || !this.atendimento.pedagogoId || this.atendimento.id === undefined) {
       this.formInvalid = true;
       return;
     }
-  
+    
     this.atendimentoService.updateAtendimento(this.atendimento.id, this.atendimento).subscribe(
       () => {
         console.log('Atendimento atualizado com sucesso.');
@@ -81,20 +103,29 @@ export class EditComponent implements OnInit {
         console.error('Erro ao atualizar o acompanhamento', error);
       }
     );
-  }  
+  }
+  
 
   deletarAtendimento(): void {
     if (confirm('Tem certeza que deseja excluir este atendimento?')) {
-      this.atendimentoService.deleteAtendimento(this.atendimento.id).subscribe(
-        () => {
-          console.log('Atendimento excluído com sucesso.');
-          alert('Atendimento excluído com sucesso.');
-          this.router.navigate(['/atendimentos/list']);
-        },
-        (error: any) => {
-          console.error('Erro ao excluir o atendimento', error);
-        }
-      );
+      if (this.atendimento.id !== undefined) { // Verifique se atendimento.id não é undefined
+        this.atendimentoService.deleteAtendimento(this.atendimento.id).subscribe(
+          () => {
+            console.log('Atendimento excluído com sucesso.');
+            alert('Atendimento excluído com sucesso.');
+            this.router.navigate(['/atendimentos/list']);
+          },
+          (error: any) => {
+            console.error('Erro ao excluir o atendimento', error);
+          }
+        );
+      } else {
+        console.error('ID do atendimento é undefined.');
+      }
     }
+  }
+  handleError(error: any) {
+    console.error('Erro na requisição:', error);
+    return throwError(error);
   }
 }
